@@ -1,47 +1,52 @@
-[tool.poetry]
+[project]
 name = "pygeosimplify"
 version = "0.0.2"
 description = "Python package for automatic, cell-based inference of clash-free simplified detector geometry."
-authors = ["Joshua Falco Beirer <jbeirer@cern.ch>"]
+authors = [
+  { name = "Joshua Falco Beirer", email = "jbeirer@cern.ch" }
+]
 repository = "https://github.com/jbeirer/pygeosimplify"
 documentation = "https://jbeirer.github.io/pygeosimplify/"
 readme = "README.md"
 packages = [
-  {include = "pygeosimplify"}
+  { include = "pygeosimplify" }
 ]
 
+requires-python = ">=3.10,<3.13"
+dynamic = ["dependencies"]
+
 [tool.poetry.dependencies]
-python = ">=3.9,<3.13"
-uproot = "^5.1.2"
-pandas = "^2.1.2"
-numpy = "^1.26.1"
-matplotlib = "^3.8.1"
-scipy = "^1.11.3"
-tqdm = "^4.66.1"
-distinctipy = "^1.2.3"
-pyg4ometry = "^1.1.1"
+python = ">=3.10,<3.13"
+uproot = "^5.5.1"
+pandas = "^2.2.3"
+numpy = "^2.2.1"
+matplotlib = "^3.10.0"
+scipy = "^1.15.1"
+tqdm = "^4.67.1"
+distinctipy = "^1.3.4"
+pyg4ometry = "^1.3.2"
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^7.2.0"
-pytest-cov = "^4.0.0"
-deptry = "^0.12.0"
-mypy = "^1.5.1"
-pre-commit = "^3.4.0"
-tox = "^4.11.1"
-ipykernel = "^6.26.0"
+pytest = "^8.3.4"
+pytest-cov = "^6.0.0"
+deptry = "^0.22.0"
+mypy = "^1.14.1"
+pre-commit = "^4.0.1"
+tox = "^4.23.2"
+ipykernel = "^6.29.5"
 
 [tool.poetry.group.docs.dependencies]
-mkdocs = "^1.4.2"
-mkdocs-material = "^9.2.7"
-mkdocstrings = {extras = ["python"], version = "^0.23.0"}
+mkdocs = "^1.6.1"
+mkdocs-material = "^9.5.49"
+mkdocstrings = {extras = ["python"], version = "^0.27.0"}
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
+requires = ["poetry-core>=2.0.1"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.black]
 line-length = 120
-target-version = ['py37']
+target-version = ['py312']
 preview = true
 
 [tool.mypy]
@@ -64,10 +69,10 @@ filterwarnings = [
 ]
 
 [tool.ruff]
-target-version = "py37"
+target-version = "py312"
 line-length = 120
 fix = true
-select = [
+lint.select = [
     # flake8-2020
     "YTT",
     # flake8-bandit
@@ -99,13 +104,19 @@ select = [
     # tryceratops
     "TRY",
 ]
-ignore = [
+lint.ignore = [
     # LineTooLong
     "E501",
     # DoNotAssignLambda
     "E731",
     # Comparison to true should be 'if cond is true:'
-    "E712"
+    "E712",
+    # Create your own exception
+    "TRY002",
+    # Avoid specifying long messages outside the exception class
+    "TRY003",
+    # B017 Do not assert blind exception: `Exception`
+    "B017"
 ]
 
 [tool.coverage.report]
@@ -116,5 +127,5 @@ branch = true
 source = ["pygeosimplify"]
 
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "tests/*" = ["S101"]
